Bug 1073081 - Fix -Wsign-compare and -Wsign-conversion warnings. r=ehsan
UltraBlame original commit: bb88a7f640faf9a43f7e965461740fd8b89cf719

+++ ipc/glue/ProcessUtils_linux.cpp
@@ -112,7 +112,7 @@ static void ProcLoaderClientDeinit();
 
 
 
-static const int kReservedFileDescriptors = 5;
+static const size_t kReservedFileDescriptors = 5;
 static const int kBeginReserveFileDescriptor = STDERR_FILENO + 1;
 
 class ProcLoaderParent : public PProcLoaderParent
